in the middle of significant restructering towards a more clear entity structure, which also supports versioned packages
