+++ src/Poseidon/SecondaryTypes.hs
@@ -6,20 +6,13 @@ module Poseidon.SecondaryTypes (
     poseidonVersionParser,
     ContributorSpec (..),
     contributorSpecParser,
-    IndividualInfo (..),
-    GroupInfo(..),
     VersionComponent (..),
     updateThreeComponentVersion,
-    PackageInfo(..),
     P.runParser,
     ORCID (..),
     ServerApiReturnType(..),
     ApiReturnData(..),
-    ExtendedIndividualInfo(..),
     processApiResponse,
-    makeNameWithVersion,
-    HasNameAndVersion (..),
-    PacNameAndVersion(..),
     ArchiveEndpoint(..),
     qDefault, qArchive
 ) where
@@ -27,6 +20,7 @@ module Poseidon.SecondaryTypes (
 import           Paths_poseidon_hs      (version)
 import           Poseidon.Utils         (PoseidonException (..), PoseidonIO,
                                          logError, logInfo)
+import Poseidon.EntityTypes (PackageInfo (..), GroupInfo (..), ExtendedIndividualInfo (..))
 
 import           Control.Exception      (catch, throwIO)
 import           Control.Monad          (forM_, guard, mzero, unless)
@@ -37,7 +31,6 @@ import           Data.Aeson             (FromJSON, ToJSON (..), Value (String),
 import           Data.Char              (digitToInt)
 import           Data.List              (intercalate)
 import           Data.Text              (pack, unpack)
-import           Data.Time              (Day)
 import           Data.Version           (Version (..), makeVersion, showVersion)
 import           GHC.Generics           (Generic)
 import           Network.HTTP.Conduit   (simpleHttp)
@@ -65,107 +58,6 @@ data ArchiveEndpoint = ArchiveEndpoint {
       _aeServerURL   :: String
     , _aeArchiveName :: Maybe String
 }
-
-class HasNameAndVersion a where
-    getPacName :: a -> String
-    getPacVersion :: a -> Maybe Version
-
-makeNameWithVersion :: (HasNameAndVersion a) => a -> String
-makeNameWithVersion a = case getPacVersion a of
-    Nothing -> getPacName a
-    Just v  -> getPacName a ++ "-" ++ showVersion v
-
-newtype PacNameAndVersion = PacNameAndVersion (String, Maybe Version) deriving (Eq, Ord)
-
-instance HasNameAndVersion PacNameAndVersion where
-    getPacName (PacNameAndVersion (n, _)) = n
-    getPacVersion (PacNameAndVersion (_, v)) = v
-
-data ExtendedIndividualInfo = ExtendedIndividualInfo
-    {
-      extIndInfoName    :: String
-    , extIndInfoGroups  :: [String]
-    , extIndInfoPacName :: String
-    , extIndInfoVersion :: Maybe Version
-    , extIndInfoAddCols :: [(String, Maybe String)]
-    }
-
-instance HasNameAndVersion ExtendedIndividualInfo where
-    getPacName = extIndInfoPacName
-    getPacVersion = extIndInfoVersion
-
-instance ToJSON ExtendedIndividualInfo where
-    toJSON e =
-        object [
-            "poseidonID" .= extIndInfoName e, -- following Janno column names
-            "groupNames" .= extIndInfoGroups e,
-            "packageTitle" .= extIndInfoPacName e, -- following mostly the Poseidon YAML definition where possible
-            "packageVersion" .= extIndInfoVersion e,
-            "additionalJannoColumns" .= extIndInfoAddCols e]
-
-instance FromJSON ExtendedIndividualInfo where
-    parseJSON = withObject "ExtendedIndividualInfo" $ \v -> ExtendedIndividualInfo
-            <$> v .: "poseidonID"
-            <*> v .: "groupNames"
-            <*> v .: "packageTitle"
-            <*> v .: "packageVersion"
-            <*> v .: "additionalJannoColumns"
-
-data PackageInfo = PackageInfo
-    { pTitle         :: String
-    , pVersion       :: Maybe Version
-    , pPosVersion    :: Version
-    , pDescription   :: Maybe String
-    , pLastModified  :: Maybe Day
-    , pNrIndividuals :: Int
-    } deriving (Eq)
-
-instance HasNameAndVersion PackageInfo where
-    getPacName = pTitle
-    getPacVersion = pVersion
-
-instance ToJSON PackageInfo where
-    toJSON (PackageInfo title pacVersion posVersion description lastModified nrIndividuals) =
-        object [
-            "packageTitle" .= title,
-            "packageVersion" .= pacVersion,
-            "poseidonVersion" .= posVersion,
-            "description" .= description,
-            "lastModified" .= lastModified,
-            "nrIndividuals" .= nrIndividuals
-        ]
-
-instance FromJSON PackageInfo where
-    parseJSON = withObject "PackageInfo" $ \v -> PackageInfo
-            <$> v .: "packageTitle"
-            <*> v .: "packageVersion"
-            <*> v .: "poseidonVersion"
-            <*> v .: "description"
-            <*> v .: "lastModified"
-            <*> v .: "nrIndividuals"
-
-data GroupInfo = GroupInfo
-    { gName          :: String
-    , gPackageNames  :: PacNameAndVersion
-    , gNrIndividuals :: Int
-    }
-
-instance ToJSON GroupInfo where
-    toJSON (GroupInfo name (PacNameAndVersion (pacTitle, pacVersion)) nrIndividuals) =
-        object [
-            "groupName" .= name,
-            "packageTitle" .= pacTitle,
-            "packageVersion" .= pacVersion,
-            "nrIndividuals" .= nrIndividuals
-        ]
-
-instance FromJSON GroupInfo where
-    parseJSON = withObject "GroupInfo" $ \v -> do
-        groupName <- v .: "groupName"
-        packageTitle <- v .: "packageTitle"
-        packageVersion <- v .: "packageVersion"
-        nrIndividuals <- v .: "nrIndividuals"
-        return $ GroupInfo groupName (PacNameAndVersion (packageTitle, packageVersion)) nrIndividuals
 
 data ServerApiReturnType = ServerApiReturnType {
     _apiMessages :: [String],
@@ -250,16 +142,6 @@ updateThreeComponentVersion component v =
             Minor -> [ i !! 0,     (i !! 1) + 1, 0          ]
             Major -> [(i !! 0) + 1,              0, 0       ]
     in makeVersion r
-
-data IndividualInfo = IndividualInfo
-    { indInfoName    :: String
-    , indInfoGroups  :: [String]
-    , indInfoPacName :: String
-    } deriving (Show, Ord, Generic)
-
-instance Eq IndividualInfo where
-    (==) (IndividualInfo a1 b1 c1) (IndividualInfo a2 b2 c2) = a1 == a2 && head b1 == head b2 && c1 == c2
-
 
 poseidonVersionParser :: P.Parser Version
 poseidonVersionParser = do
